Basic MPC Working in Simulink Using CasADi
The horizon length is fixed so it think it always is at the start of a maneuver making the performance not great. Also, does not converge very well when the rocket rotates. Both these issues will hopefully be fixed by adding the final time as an optimization parameter. Added a small visualization script too