Added Test Assessments, Updated Controllers with Variable Final State
The delay tests now check if the sim stopped because the error was low enough instead of crashing into the ground or not being able to solve the optimization problem. Also, added scenarios using the signal editor and test sequence blocks. Cleaned up the Simulink model. Started work on adding state estimation.